[FormBulder]
Extensive update to the FormBuilder. Bumped C1 requirement to 4.0.

Removed registration of IData related features since its no near ready and focused on stabilizing the Code First
approach which is now considered stable to use in production.

The ajax-based validation is known to have some issues and will be looked at in the next service release.