Got Page Preview to work by resolving the route for the page, executing it and injecting the resulting content as a Literal Control on renderers/page.aspx.
Refactored Page Templates so we now use Strong Models and the same attributes as normal C1 templates for defining Content Placeholders.
 - A pages template is exposed as ViewBag.TemplateModel and its placeholder-properties can be obtained this way.
 - So instead of Html.C1().TemplatePlaceHolder("Content") its now ViewBag.TemplateModel.Content

Added a RenderingReason contstraint for routes, so different actions can be specified for ie. PageView and PreviewUnsavedChanges.

Made MvcFunctions work on Non Mvc templates by executing the Controller as a direct route. Limitations to this are
 - There is no parent controller to the function since its executed standalone/byitself
 - Actions can't have the ChildActionOnly-attribute
 - These limitations of course doesn't apply if the Page template is a Mvc Controller.