Updated the ScheduledTasks package to work with HangFire 1.5.
Removed dependency on Hangfire.CompositeC1 so we can use any JobStorage we want.

Added a new ScheduledTasksConfiguration class which adds support for aggregating
IBackgroundProcess instances to use in Hangfire. IBackgroundProcess is a new
interface added in Hangfire 1.5 which makes it easy to have infine-running
processes running inside the Hangfire infrastructure instead of handling own
threads.