Bigger update to ECommerce with breaking changes
New PaymentRequest concept introduced, where one has to explicitly create a
PaymentRequest to an order to initiate a payment flow. When creating a request
its possible to define which provider to use which finally means that we can
operate with more than one default provider.

In the scenarios where an order doesn't need payment, its not required to
create such a request but the user can be redirected directly to the
Continue Url for the order.

When adding multiple providers, its also possible to inherit all non-specified
values from a parent provider - this is convenient when one need multiple
providers for different countries/markets, but all that differs is ie.
Payment Methods. By inheriting from a base provider, one is not needed to
duplicate keys and other basic configuration data.